requirements.txt
langchain
langchain-text-splitters
langchain-community
langchain_openai
langchain-chroma
chromadb
openai
python-dotenv
pypdf